Add a simple test for suspense
